Some adjustments around many place.

+++ src/main/java/big/company/domain/Employee.java
@@ -1,3 +1,8 @@
 package big.company.domain;
 
-public record Employee(int id, String firstName, String lastName, int salary, Integer managerId) {}
+import java.math.BigDecimal;
+
+/**
+ * Holds the data of an employee.
+ */
+public record Employee(int id, String firstName, String lastName, BigDecimal salary, Integer managerId) {}
