+++ src/main/java/big/company/domain/EmployeeStatistics.java
@@ -1,7 +1,11 @@
 package big.company.domain;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
-public record EmployeeStatistics(Map<Employee, Double> managersWithLessExpectedEarning,
-                                 Map<Employee, Double> managersWithMoreExpectedEarning,
+/**
+ * Holds the data of the calculated statistics.
+ */
+public record EmployeeStatistics(Map<Employee, BigDecimal> managersWithLessExpectedEarning,
+                                 Map<Employee, BigDecimal> managersWithMoreExpectedEarning,
                                  Map<Employee, Integer> employeesWithLongReportingLine) {}
