+++ src/main/java/big/company/domain/ManagersWithEarningDifferenceMapContainer.java
@@ -1,6 +1,10 @@
 package big.company.domain;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
-public record ManagersWithEarningDifferenceMapContainer(Map<Employee, Double> managersWithLessExpectedEarning,
-                                                        Map<Employee, Double> managersWithMoreExpectedEarning) {}
+/**
+ * Holds data for managers earning.
+ */
+public record ManagersWithEarningDifferenceMapContainer(Map<Employee, BigDecimal> managersWithLessExpectedEarning,
+                                                        Map<Employee, BigDecimal> managersWithMoreExpectedEarning) {}
